requirements.txt
python-gitlab==1.2.0
